+++ cliquet/tests/test_views_batch.py
@@ -46,7 +46,7 @@ class BatchViewTest(BaseWebTest, unittest.TestCase):
         hello = resp.json['responses'][0]
         self.assertEqual(hello['path'], '/v0/')
         self.assertEqual(hello['status'], 200)
-        self.assertEqual(hello['body']['hello'], 'myapp')
+        self.assertEqual(hello['body']['project_name'], 'myapp')
         self.assertIn('application/json', hello['headers']['Content-Type'])
 
     def test_empty_response_body_with_head(self):
@@ -88,7 +88,7 @@ class BatchViewTest(BaseWebTest, unittest.TestCase):
         hello = resp.json['responses'][0]
         self.assertEqual(hello['path'], '/v0/')
         self.assertEqual(hello['status'], 200)
-        self.assertEqual(hello['body']['hello'], 'myapp')
+        self.assertEqual(hello['body']['project_name'], 'myapp')
         self.assertIn('application/json', hello['headers']['Content-Type'])
 
     def test_redirect_responses_are_followed(self):